Update pvl_perez.m (#24)
* Update pvl_perez.m

These correlations are not numerically stable; i.e. you can get physically unreasonable values for inputs that are reasonable (I'll send figure separately to Cliff). And unreasonable inputs that generate unreasonable outputs are not flagged. This change warns user of unreasonable outputs and converts them to NaNs so unreasonable values are not propagated.

* Update pvl_perez.m

Added comments. Thanks.

* Update pvl_perez.m